requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4     
annotated-types==0.7.0   
anyio==4.12.1
audioop-lts==0.2.2       
brotli==1.2.0
certifi==2026.1.4        
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.129.0
ffmpy==1.0.0
filelock==3.24.3
fonttools==4.61.1        
fsspec==2026.2.0
gradio==6.6.0
gradio_client==2.1.0     
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
nvidia-ml-py==13.590.48
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
reportlab==4.4.10
requests==2.32.5
rich==14.3.3
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
segmentation_models_pytorch==0.5.0
semantic-version==2.10.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.24
tomlkit==0.13.3
torch==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.3
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0